+++ src/index.js
@@ -21,7 +21,7 @@ export async function onLoadAnotherPage(e) {
 }
 import './js/movie_search';
 import { studentCards } from './js/footer-modal';
-import './js/my-library-watched-queue/renderMoviesToLibrary';
+import './js/my-library-watched-queue/my-library-watched-queue';
 // import Darkmode from 'darkmode-js';
 // //https://darkmodejs.learn.uno
