+++ src/js/my-library-watched-queue/my-library-watched-queue.js
@@ -1,0 +1,44 @@
+import Notiflix from 'notiflix';
+import { renderMoviesLibrary } from './renderMoviesLibrary';
+import { userDataQueue, userDataWatched } from './data-template-local-storage';
+
+const watchedBtnEl = document.querySelector('.js-library-btn--watched');
+const queueBtnEl = document.querySelector('.js-library-btn--queue');
+const movieListEl = document.querySelector('.movie-list');
+
+// при натисканні на Watched
+function onWatchedBtnClick() {
+  const movieWatched = gettingItem('movieWatched');
+  if (movieWatched.length === 0) {
+    movieListEl.innerHTML = '<p>Ви ще не переглянули жодного фільму</p>';
+    return;
+  }
+  const moviesCards = movieWatched
+    .map(movie => {
+      return renderMoviesLibrary(movie);
+    })
+    .join('');
+  movieListEl.innerHTML = moviesCards;
+  console.log(movieWatched);
+}
+
+watchedBtnEl.addEventListener('click', onWatchedBtnClick);
+// queueBtnEl.addEventListener('click', onQueueBtnClick);
+
+// localStorage functions
+function gettingItem(key) {
+  try {
+    const serializedState = localStorage.getItem(key);
+    return serializedState === null ? undefined : JSON.parse(serializedState);
+  } catch (error) {
+    console.log('Everyone makes mistakes, this is yours:', error.message);
+  }
+}
+
+// function removingItem(key) {
+//   try {
+//     localStorage.removeItem(key);
+//   } catch (error) {
+//     console.log('Everyone makes mistakes, this is yours:', error.message);
+//   }
+// }
