+++ src/library.html
@@ -20,7 +20,7 @@
     <script type="module" src="./index.js"></script>
     <script
       type="module"
-      src="./js/my-library-watched-queue/renderMoviesToLibrary.js"
+      src="./js/my-library-watched-queue/my-library-watched-queue.js"
     ></script>
   </head>
   <body>
