+++ src/sass/components/_movie_card.scss
@@ -1,14 +1,92 @@
 .movie-rating {
-  padding-right: 10px;
-  padding-left: 10px;
+  color: $primary-white-color;
+  border-radius: 5px;
+  background-color: $main-accent-color;
   width: 36px;
   height: 16px;
-
+  padding-right: 10px;
+  padding-left: 10px;
   font-weight: 500;
   font-size: 12px;
   line-height: 1.17em;
-  color: $primary-white-color;
 
+  @media screen and (min-width: 1280px) {
+    width: 40px;
+    height: 20px;
+    padding-right: 12px;
+    padding-left: 12px;
+  }
+}
+
+.movie-list {
+  @media screen and (min-width: 768px) {
+    display: flex;
+    flex-wrap: wrap;
+    column-gap: 32px;
+    row-gap: 24px;
+  }
+
+  @media screen and (min-width: 1280px) {
+    column-gap: 16px;
+    row-gap: 20px;
+  }
+
+  &__title {
+    font-weight: 500;
+    font-size: 12px;
+    line-height: 1.33;
+    text-transform: uppercase;
+
+    @media screen and (min-width: 768px) {
+      font-weight: 500;
+      font-size: 12px;
+    }
+
+    @media screen and (min-width: 1280px) {
+      font-style: normal;
+      font-weight: 500;
+      font-size: 20px;
+      color: rgba(0, 0, 0, 1);
+    }
+  }
+
+  &__info {
+    font-style: normal;
+    font-weight: 500;
+    font-size: 12px;
+    line-height: 1.33;
+    text-transform: uppercase;
+    color: rgba(255, 107, 8, 1);
+
+    @media screen and (min-width: 768px) {
+      font-size: 12px;
+    }
+
+    @media screen and (min-width: 1280px) {
+      font-weight: 500;
+      font-size: 20px;
+    }
+  }
+}
+
+.library-gallery__image {
+  width: 280px;
+  height: 402px;
   border-radius: 5px;
-  background-color: $main-accent-color;
+  object-fit: cover;
+
+  @media screen and (min-width: 320px) {
+    width: 280px;
+    height: 402px;
+  }
+
+  @media screen and (min-width: 768px) {
+    width: 336px;
+    height: 455px;
+  }
+
+  @media screen and (min-width: 1280px) {
+    width: 394px;
+    height: 574px;
+  }
 }
